moved functionality away from test_bs to setup_utils mainly
