+++ tests/block_atx/add_node/test_blocks_add_node.py
@@ -1,8 +1,8 @@
 from pytest_testconfig import config as testconfig
 
 from tests import queries as q
+from tests.setup_utils import add_multi_clients
 from tests.test_bs import setup_network, add_curl, setup_bootstrap, start_poet, setup_clients, wait_genesis
-from tests.test_bs import add_multi_clients
 from tests.utils import validate_blocks_per_nodes, get_pod_id, get_conf
 
 
@@ -45,7 +45,7 @@ def test_add_node_validate_atx(init_session, setup_network):
     print("adding a new miner")
     bs_info = setup_network.bootstrap.pods[0]
     cspec = get_conf(bs_info, testconfig['client'], testconfig['genesis_delta'])
-    new_pod_name = add_multi_clients(init_session, cspec, 1)[0]
+    new_pod_name = add_multi_clients(testconfig, init_session, cspec, 1)[0]
 
     # wait for next epoch
     last_layer = layers_per_epoch * (curr_epoch + 1)
