+++ tests/convenience.py
@@ -5,6 +5,8 @@ TIMESTAMP_FMT = "%Y-%m-%dT%H:%M:%S.%fZ"
 
 PRINT_SEP = "===================================================================="
 
+LOG_ENTRIES = {"message": "M"}
+
 
 def print_nl_end(*args):
     print(*args, "\n")
@@ -24,3 +26,29 @@ def sleep_print_backwards(tts):
         tts -= 1
         print(f" {tts} seconds left     ", end="\r")
         time.sleep(1)
+
+
+def print_hits_entry_count(hits, log_entry):
+    """
+    prints number of seen values
+
+    :param hits: list, all query results
+    :param log_entry: string, what part of the log do we want to sum
+    """
+
+    if log_entry not in LOG_ENTRIES.keys():
+        raise ValueError(f"unknown log entry {log_entry}")
+
+    result = {}
+    entry = LOG_ENTRIES[log_entry]
+
+    for hit in hits:
+        entry_val = getattr(hit, entry)
+        if entry_val not in result:
+            result[entry_val] = 1
+            continue
+
+        result[entry_val] += 1
+
+    for key in result:
+        print(f"found {result[key]} appearances of '{key}' in hits")
