+++ tests/stress/grpc_stress/test_stress_grpc.py
@@ -3,14 +3,13 @@ from pytest_testconfig import config as testconfig
 import multiprocessing as mp
 import random
 
-from tests.convenience import sleep_print_backwards
+from tests.convenience import sleep_print_backwards, print_hits_entry_count
 from tests.queries import find_error_log_msgs
 from tests.test_bs import setup_network, add_curl, setup_bootstrap, start_poet, setup_clients, wait_genesis
 from tests.tx_generator import actions
 from tests.tx_generator import config as conf
 from tests.tx_generator.models.wallet_api import WalletAPI
 from tests.tx_generator.models.accountant import Accountant
-from tests.utils import print_hits_entry_count
 
 
 def test_grpc_stress(init_session, setup_network):
