+++ tests/stress/sync_stress/test_sync.py
@@ -2,9 +2,10 @@ from datetime import datetime
 import time
 from pytest_testconfig import config as testconfig
 
-from tests.test_bs import add_multi_clients, setup_bootstrap, save_log_on_exit
-import tests.queries as q
 from tests.convenience import convert_ts_to_datetime
+import tests.queries as q
+from tests.setup_utils import add_multi_clients
+from tests.test_bs import setup_bootstrap, save_log_on_exit
 from tests.utils import get_conf
 
 # ==============================================================================
@@ -23,7 +24,7 @@ def test_sync_stress(init_session, setup_bootstrap, save_log_on_exit):
     clients_num = testconfig["client"]["replicas"]
     bs_info = setup_bootstrap.pods[0]
     cspec = get_conf(bs_info, testconfig['client'], testconfig['genesis_delta'])
-    _ = add_multi_clients(init_session, cspec, clients_num)
+    _ = add_multi_clients(testconfig, init_session, cspec, clients_num)
 
     hits = []
     number_of_pods = clients_num + 1  # add 1 for bootstrap pod
@@ -37,7 +38,7 @@ def test_sync_stress(init_session, setup_bootstrap, save_log_on_exit):
     cspec.args['data-folder'] = ""
 
     # Adding a single new client
-    res_lst = add_multi_clients(init_session, cspec, 1, 'client')
+    res_lst = add_multi_clients(testconfig, init_session, cspec, 1, 'client')
     new_client = res_lst[0]
 
     # wait for the new node to start syncing
